requirements.txt
anyio==4.6.2.post1
argcomplete==3.5.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.11.0
blivet-gui==2.5.0
boto3==1.35.55
botocore==1.35.55
Brlapi==0.8.5
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.21.1
dnspython==2.7.0
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.9.0
idna==3.7
iso639==0.1.4
jmespath==1.0.1
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.3
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
MarkupSafe==2.1.5
nftables==0.1
olefile==0.47
packaging==24.1
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
ply==3.11
productmd==1.41
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.48.2
pykickstart==3.58
pymongo==4.10.1
pyOpenSSL==24.2.1
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-telegram-bot==21.6
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.0
s3transfer==0.10.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.8.1
soupsieve==2.6
systemd-python==235
Tempita==0.5.2
urllib3==1.26.20
xkbregistry==0.3
